chore: Move extensions/pkg/predicate to pkg/controllerutils/predicate
