+++ extensions/pkg/controller/extension/controller.go
@@ -19,6 +19,7 @@ import (
 	extensionscontroller "github.com/gardener/gardener/extensions/pkg/controller"
 	extensionspredicate "github.com/gardener/gardener/extensions/pkg/predicate"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	predicateutils "github.com/gardener/gardener/pkg/controllerutils/predicate"
 )
 
 const (
@@ -65,8 +66,8 @@ func DefaultPredicates(ctx context.Context, mgr manager.Manager, ignoreOperation
 }
 
 func add(mgr manager.Manager, args AddArgs) error {
-	predicates := []predicate.Predicate{extensionspredicate.HasType(args.Type)}
-	predicates = append(predicates, extensionspredicate.HasClass(args.ExtensionClasses...))
+	predicates := []predicate.Predicate{predicateutils.HasType(args.Type)}
+	predicates = append(predicates, predicateutils.HasClass(args.ExtensionClasses...))
 	predicates = append(predicates, args.Predicates...)
 
 	c, err := builder.
