--- extensions/pkg/predicate/predicate_suite_test.go
@@ -1,17 +1,0 @@
-// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package predicate_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestPredicate(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Extensions Predicate Suite")
-}
