+++ pkg/controllerutils/predicate/predicate_test.go
@@ -5,10 +5,13 @@
 package predicate_test
 
 import (
+	"reflect"
+
 	. "github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
+	. "github.com/onsi/gomega"
 	gomegatypes "github.com/onsi/gomega/types"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
@@ -20,6 +23,10 @@ import (
 )
 
 var _ = Describe("Predicate", func() {
+	var (
+		extensionType = "extension-type"
+	)
+
 	Describe("#IsDeleting", func() {
 		var (
 			shoot        *gardencorev1beta1.Shoot
@@ -54,10 +61,10 @@ var _ = Describe("Predicate", func() {
 
 		Context("shoot doesn't have a deletion timestamp", func() {
 			It("should be false", func() {
-				gomega.Expect(predicate.Create(createEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Update(updateEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Delete(deleteEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Generic(genericEvent)).To(gomega.BeFalse())
+				Expect(predicate.Create(createEvent)).To(BeFalse())
+				Expect(predicate.Update(updateEvent)).To(BeFalse())
+				Expect(predicate.Delete(deleteEvent)).To(BeFalse())
+				Expect(predicate.Generic(genericEvent)).To(BeFalse())
 			})
 		})
 
@@ -69,10 +76,10 @@ var _ = Describe("Predicate", func() {
 			})
 
 			It("should be true", func() {
-				gomega.Expect(predicate.Create(createEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Update(updateEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Delete(deleteEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Generic(genericEvent)).To(gomega.BeTrue())
+				Expect(predicate.Create(createEvent)).To(BeTrue())
+				Expect(predicate.Update(updateEvent)).To(BeTrue())
+				Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+				Expect(predicate.Generic(genericEvent)).To(BeTrue())
 			})
 		})
 	})
@@ -111,10 +118,10 @@ var _ = Describe("Predicate", func() {
 
 		Context("shoot has the requested name", func() {
 			It("should be true", func() {
-				gomega.Expect(predicate.Create(createEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Update(updateEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Delete(deleteEvent)).To(gomega.BeTrue())
-				gomega.Expect(predicate.Generic(genericEvent)).To(gomega.BeTrue())
+				Expect(predicate.Create(createEvent)).To(BeTrue())
+				Expect(predicate.Update(updateEvent)).To(BeTrue())
+				Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+				Expect(predicate.Generic(genericEvent)).To(BeTrue())
 			})
 		})
 
@@ -124,10 +131,10 @@ var _ = Describe("Predicate", func() {
 			})
 
 			It("should be false", func() {
-				gomega.Expect(predicate.Create(createEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Update(updateEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Delete(deleteEvent)).To(gomega.BeFalse())
-				gomega.Expect(predicate.Generic(genericEvent)).To(gomega.BeFalse())
+				Expect(predicate.Create(createEvent)).To(BeFalse())
+				Expect(predicate.Update(updateEvent)).To(BeFalse())
+				Expect(predicate.Delete(deleteEvent)).To(BeFalse())
+				Expect(predicate.Generic(genericEvent)).To(BeFalse())
 			})
 		})
 	})
@@ -136,28 +143,28 @@ var _ = Describe("Predicate", func() {
 		func(events []EventType, createMatcher, updateMatcher, deleteMatcher, genericMatcher gomegatypes.GomegaMatcher) {
 			p := ForEventTypes(events...)
 
-			gomega.Expect(p.Create(event.CreateEvent{})).To(createMatcher)
-			gomega.Expect(p.Update(event.UpdateEvent{})).To(updateMatcher)
-			gomega.Expect(p.Delete(event.DeleteEvent{})).To(deleteMatcher)
-			gomega.Expect(p.Generic(event.GenericEvent{})).To(genericMatcher)
+			Expect(p.Create(event.CreateEvent{})).To(createMatcher)
+			Expect(p.Update(event.UpdateEvent{})).To(updateMatcher)
+			Expect(p.Delete(event.DeleteEvent{})).To(deleteMatcher)
+			Expect(p.Generic(event.GenericEvent{})).To(genericMatcher)
 		},
 
-		Entry("none", nil, gomega.BeFalse(), gomega.BeFalse(), gomega.BeFalse(), gomega.BeFalse()),
-		Entry("create", []EventType{Create}, gomega.BeTrue(), gomega.BeFalse(), gomega.BeFalse(), gomega.BeFalse()),
-		Entry("update", []EventType{Update}, gomega.BeFalse(), gomega.BeTrue(), gomega.BeFalse(), gomega.BeFalse()),
-		Entry("delete", []EventType{Delete}, gomega.BeFalse(), gomega.BeFalse(), gomega.BeTrue(), gomega.BeFalse()),
-		Entry("generic", []EventType{Generic}, gomega.BeFalse(), gomega.BeFalse(), gomega.BeFalse(), gomega.BeTrue()),
-		Entry("create, update", []EventType{Create, Update}, gomega.BeTrue(), gomega.BeTrue(), gomega.BeFalse(), gomega.BeFalse()),
-		Entry("create, delete", []EventType{Create, Delete}, gomega.BeTrue(), gomega.BeFalse(), gomega.BeTrue(), gomega.BeFalse()),
-		Entry("create, generic", []EventType{Create, Generic}, gomega.BeTrue(), gomega.BeFalse(), gomega.BeFalse(), gomega.BeTrue()),
-		Entry("update, delete", []EventType{Update, Delete}, gomega.BeFalse(), gomega.BeTrue(), gomega.BeTrue(), gomega.BeFalse()),
-		Entry("update, generic", []EventType{Update, Generic}, gomega.BeFalse(), gomega.BeTrue(), gomega.BeFalse(), gomega.BeTrue()),
-		Entry("delete, generic", []EventType{Delete, Generic}, gomega.BeFalse(), gomega.BeFalse(), gomega.BeTrue(), gomega.BeTrue()),
-		Entry("create, update, delete", []EventType{Create, Update, Delete}, gomega.BeTrue(), gomega.BeTrue(), gomega.BeTrue(), gomega.BeFalse()),
-		Entry("create, update, generic", []EventType{Create, Update, Generic}, gomega.BeTrue(), gomega.BeTrue(), gomega.BeFalse(), gomega.BeTrue()),
-		Entry("create, delete, generic", []EventType{Create, Delete, Generic}, gomega.BeTrue(), gomega.BeFalse(), gomega.BeTrue(), gomega.BeTrue()),
-		Entry("update, delete, generic", []EventType{Update, Delete, Generic}, gomega.BeFalse(), gomega.BeTrue(), gomega.BeTrue(), gomega.BeTrue()),
-		Entry("create, update, delete, generic", []EventType{Create, Update, Delete, Generic}, gomega.BeTrue(), gomega.BeTrue(), gomega.BeTrue(), gomega.BeTrue()),
+		Entry("none", nil, BeFalse(), BeFalse(), BeFalse(), BeFalse()),
+		Entry("create", []EventType{Create}, BeTrue(), BeFalse(), BeFalse(), BeFalse()),
+		Entry("update", []EventType{Update}, BeFalse(), BeTrue(), BeFalse(), BeFalse()),
+		Entry("delete", []EventType{Delete}, BeFalse(), BeFalse(), BeTrue(), BeFalse()),
+		Entry("generic", []EventType{Generic}, BeFalse(), BeFalse(), BeFalse(), BeTrue()),
+		Entry("create, update", []EventType{Create, Update}, BeTrue(), BeTrue(), BeFalse(), BeFalse()),
+		Entry("create, delete", []EventType{Create, Delete}, BeTrue(), BeFalse(), BeTrue(), BeFalse()),
+		Entry("create, generic", []EventType{Create, Generic}, BeTrue(), BeFalse(), BeFalse(), BeTrue()),
+		Entry("update, delete", []EventType{Update, Delete}, BeFalse(), BeTrue(), BeTrue(), BeFalse()),
+		Entry("update, generic", []EventType{Update, Generic}, BeFalse(), BeTrue(), BeFalse(), BeTrue()),
+		Entry("delete, generic", []EventType{Delete, Generic}, BeFalse(), BeFalse(), BeTrue(), BeTrue()),
+		Entry("create, update, delete", []EventType{Create, Update, Delete}, BeTrue(), BeTrue(), BeTrue(), BeFalse()),
+		Entry("create, update, generic", []EventType{Create, Update, Generic}, BeTrue(), BeTrue(), BeFalse(), BeTrue()),
+		Entry("create, delete, generic", []EventType{Create, Delete, Generic}, BeTrue(), BeFalse(), BeTrue(), BeTrue()),
+		Entry("update, delete, generic", []EventType{Update, Delete, Generic}, BeFalse(), BeTrue(), BeTrue(), BeTrue()),
+		Entry("create, update, delete, generic", []EventType{Create, Update, Delete, Generic}, BeTrue(), BeTrue(), BeTrue(), BeTrue()),
 	)
 
 	Describe("#RelevantConditionsChanged", func() {
@@ -179,48 +186,48 @@ var _ = Describe("Predicate", func() {
 
 		Describe("#Create", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Create(event.CreateEvent{})).To(gomega.BeTrue())
+				Expect(p.Create(event.CreateEvent{})).To(BeTrue())
 			})
 		})
 
 		Describe("#Update", func() {
 			It("should return false because there is no relevant change", func() {
-				gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: shoot})).To(gomega.BeFalse())
+				Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: shoot})).To(BeFalse())
 			})
 
 			tests := func(conditionType gardencorev1beta1.ConditionType) {
 				It("should return true because condition was added", func() {
 					oldShoot := shoot.DeepCopy()
 					shoot.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition was removed", func() {
 					shoot.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := shoot.DeepCopy()
 					shoot.Status.Conditions = nil
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition status was changed", func() {
 					shoot.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := shoot.DeepCopy()
 					shoot.Status.Conditions[0].Status = gardencorev1beta1.ConditionTrue
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition reason was changed", func() {
 					shoot.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := shoot.DeepCopy()
 					shoot.Status.Conditions[0].Reason = "reason"
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition message was changed", func() {
 					shoot.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := shoot.DeepCopy()
 					shoot.Status.Conditions[0].Message = "message"
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: shoot, ObjectOld: oldShoot})).To(BeTrue())
 				})
 			}
 
@@ -235,13 +242,13 @@ var _ = Describe("Predicate", func() {
 
 		Describe("#Delete", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Delete(event.DeleteEvent{})).To(gomega.BeTrue())
+				Expect(p.Delete(event.DeleteEvent{})).To(BeTrue())
 			})
 		})
 
 		Describe("#Generic", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Generic(event.GenericEvent{})).To(gomega.BeTrue())
+				Expect(p.Generic(event.GenericEvent{})).To(BeTrue())
 			})
 		})
 	})
@@ -259,48 +266,48 @@ var _ = Describe("Predicate", func() {
 
 		Describe("#Create", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Create(event.CreateEvent{})).To(gomega.BeTrue())
+				Expect(p.Create(event.CreateEvent{})).To(BeTrue())
 			})
 		})
 
 		Describe("#Update", func() {
 			It("should return false because there is no relevant change", func() {
-				gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: managedResource})).To(gomega.BeFalse())
+				Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: managedResource})).To(BeFalse())
 			})
 
 			tests := func(conditionType gardencorev1beta1.ConditionType) {
 				It("should return true because condition was added", func() {
 					oldShoot := managedResource.DeepCopy()
 					managedResource.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition was removed", func() {
 					managedResource.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := managedResource.DeepCopy()
 					managedResource.Status.Conditions = nil
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition status was changed", func() {
 					managedResource.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := managedResource.DeepCopy()
 					managedResource.Status.Conditions[0].Status = gardencorev1beta1.ConditionTrue
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition reason was changed", func() {
 					managedResource.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := managedResource.DeepCopy()
 					managedResource.Status.Conditions[0].Reason = "reason"
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(BeTrue())
 				})
 
 				It("should return true because condition message was changed", func() {
 					managedResource.Status.Conditions = []gardencorev1beta1.Condition{{Type: conditionType}}
 					oldShoot := managedResource.DeepCopy()
 					managedResource.Status.Conditions[0].Message = "message"
-					gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(gomega.BeTrue())
+					Expect(p.Update(event.UpdateEvent{ObjectNew: managedResource, ObjectOld: oldShoot})).To(BeTrue())
 				})
 			}
 
@@ -319,13 +326,13 @@ var _ = Describe("Predicate", func() {
 
 		Describe("#Delete", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Delete(event.DeleteEvent{})).To(gomega.BeTrue())
+				Expect(p.Delete(event.DeleteEvent{})).To(BeTrue())
 			})
 		})
 
 		Describe("#Generic", func() {
 			It("should return true", func() {
-				gomega.Expect(p.Generic(event.GenericEvent{})).To(gomega.BeTrue())
+				Expect(p.Generic(event.GenericEvent{})).To(BeTrue())
 			})
 		})
 	})
@@ -356,28 +363,28 @@ var _ = Describe("Predicate", func() {
 		It("should return false for all events because the extension backupbucket has operation annotation reconcile", func() {
 			metav1.SetMetaDataAnnotation(&extensionBackupBucket.ObjectMeta, "gardener.cloud/operation", "reconcile")
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Delete(event.DeleteEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Generic(event.GenericEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Delete(event.DeleteEvent{Object: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Generic(event.GenericEvent{Object: extensionBackupBucket})).To(BeFalse())
 		})
 
 		It("should not return false for create events just because the extension backupEntry has operation annotation restore or migrate", func() {
 			extensionBackupEntry.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateFailed}
 			metav1.SetMetaDataAnnotation(&extensionBackupEntry.ObjectMeta, "gardener.cloud/operation", "migrate")
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupEntry})).To(gomega.BeTrue())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupEntry})).To(BeTrue())
 
 			metav1.SetMetaDataAnnotation(&extensionBackupEntry.ObjectMeta, "gardener.cloud/operation", "restore")
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupEntry})).To(gomega.BeTrue())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupEntry})).To(BeTrue())
 		})
 
 		It("should return false for update because the extension backupEntry has operation annotation restore but the old backupEntry doesn't have it", func() {
 			oldExtensionBackupEntry := extensionBackupEntry.DeepCopy()
 			metav1.SetMetaDataAnnotation(&extensionBackupEntry.ObjectMeta, "gardener.cloud/operation", "restore")
 
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(gomega.BeFalse())
+			Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(BeFalse())
 		})
 
 		It("should not return false for update because of the operation annotation restore or migrate when the old backupEntry also have it", func() {
@@ -385,34 +392,34 @@ var _ = Describe("Predicate", func() {
 			oldExtensionBackupEntry := extensionBackupEntry.DeepCopy()
 			extensionBackupEntry.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateSucceeded}
 
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(BeTrue())
 
 			metav1.SetMetaDataAnnotation(&extensionBackupEntry.ObjectMeta, "gardener.cloud/operation", "migrate")
 			metav1.SetMetaDataAnnotation(&oldExtensionBackupEntry.ObjectMeta, "gardener.cloud/operation", "migrate")
 
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectOld: oldExtensionBackupEntry, ObjectNew: extensionBackupEntry})).To(BeTrue())
 		})
 
 		It("should return false for create and update because the extension backupbucket status has no lastOperation present", func() {
 			extensionBackupBucket.Status.LastOperation = nil
 			newExtensionBackupBucket := extensionBackupBucket.DeepCopy()
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Delete(event.DeleteEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Generic(event.GenericEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Delete(event.DeleteEvent{Object: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Generic(event.GenericEvent{Object: extensionBackupBucket})).To(BeFalse())
 		})
 
 		It("should return true for create events because the extension backupbucket status lastOperation state is Failed", func() {
 			extensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateFailed}
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(BeTrue())
 		})
 
 		It("should return false for create events because the extension backupbucket status lastOperation state is not Failed", func() {
 			extensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateSucceeded}
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(gomega.BeFalse())
+			Expect(p.Create(event.CreateEvent{Object: extensionBackupBucket})).To(BeFalse())
 		})
 
 		It("should return true for  update events because the extension backupbucket status lastOperation state is Succeeded or Error or Failed and the old state is Processing", func() {
@@ -420,13 +427,13 @@ var _ = Describe("Predicate", func() {
 			newExtensionBackupBucket := extensionBackupBucket.DeepCopy()
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateSucceeded}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateError}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateFailed}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 		})
 
 		It("should return true for update events because the extension backupbucket status lastOperation state is Succeeded or Error or Failed and the old state is nil", func() {
@@ -434,13 +441,13 @@ var _ = Describe("Predicate", func() {
 			newExtensionBackupBucket := extensionBackupBucket.DeepCopy()
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateSucceeded}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateError}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateFailed}
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeTrue())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeTrue())
 		})
 
 		It("should return false for update events because the extension backupbucket status lastOperation has changed from Succeeded or Error to Processing", func() {
@@ -448,17 +455,17 @@ var _ = Describe("Predicate", func() {
 			newExtensionBackupBucket := extensionBackupBucket.DeepCopy()
 			newExtensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateProcessing}
 
-			gomega.Expect(p.Create(event.CreateEvent{Object: newExtensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Delete(event.DeleteEvent{Object: newExtensionBackupBucket})).To(gomega.BeFalse())
-			gomega.Expect(p.Generic(event.GenericEvent{Object: newExtensionBackupBucket})).To(gomega.BeFalse())
+			Expect(p.Create(event.CreateEvent{Object: newExtensionBackupBucket})).To(BeFalse())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeFalse())
+			Expect(p.Delete(event.DeleteEvent{Object: newExtensionBackupBucket})).To(BeFalse())
+			Expect(p.Generic(event.GenericEvent{Object: newExtensionBackupBucket})).To(BeFalse())
 		})
 
 		It("should return false for update events because the extension backupbucket status lastOperation is Succeeded but it's same as old Object", func() {
 			extensionBackupBucket.Status.LastOperation = &gardencorev1beta1.LastOperation{State: gardencorev1beta1.LastOperationStateSucceeded}
 			newExtensionBackupBucket := extensionBackupBucket.DeepCopy()
 
-			gomega.Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(gomega.BeFalse())
+			Expect(p.Update(event.UpdateEvent{ObjectNew: newExtensionBackupBucket, ObjectOld: extensionBackupBucket})).To(BeFalse())
 		})
 	})
 
@@ -471,21 +478,21 @@ var _ = Describe("Predicate", func() {
 
 		It("should return false because last operation is nil on new object", func() {
 			oldLastOperation := lastOperation.DeepCopy()
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeFalse())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeFalse())
 		})
 
 		It("should return false because last operation type is 'Delete' on old object", func() {
 			lastOperation = &gardencorev1beta1.LastOperation{}
 			oldLastOperation := lastOperation.DeepCopy()
 			oldLastOperation.Type = gardencorev1beta1.LastOperationTypeDelete
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeFalse())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeFalse())
 		})
 
 		It("should return false because last operation type is 'Delete' on new object", func() {
 			lastOperation = &gardencorev1beta1.LastOperation{}
 			lastOperation.Type = gardencorev1beta1.LastOperationTypeDelete
 			oldLastOperation := lastOperation.DeepCopy()
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeFalse())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeFalse())
 		})
 
 		It("should return false because last operation type is not 'Processing' on old object", func() {
@@ -493,7 +500,7 @@ var _ = Describe("Predicate", func() {
 			lastOperation.Type = gardencorev1beta1.LastOperationTypeReconcile
 			lastOperation.State = gardencorev1beta1.LastOperationStateSucceeded
 			oldLastOperation := lastOperation.DeepCopy()
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeFalse())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeFalse())
 		})
 
 		It("should return false because last operation type is not 'Succeeded' on new object", func() {
@@ -502,7 +509,7 @@ var _ = Describe("Predicate", func() {
 			lastOperation.State = gardencorev1beta1.LastOperationStateProcessing
 			oldLastOperation := lastOperation.DeepCopy()
 			oldLastOperation.State = gardencorev1beta1.LastOperationStateProcessing
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeFalse())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeFalse())
 		})
 
 		It("should return true because last operation type is 'Succeeded' on new object", func() {
@@ -511,7 +518,351 @@ var _ = Describe("Predicate", func() {
 			lastOperation.State = gardencorev1beta1.LastOperationStateSucceeded
 			oldLastOperation := lastOperation.DeepCopy()
 			oldLastOperation.State = gardencorev1beta1.LastOperationStateProcessing
-			gomega.Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(gomega.BeTrue())
+			Expect(ReconciliationFinishedSuccessfully(oldLastOperation, lastOperation)).To(BeTrue())
+		})
+	})
+
+	Describe("#HasType", func() {
+		var (
+			object       client.Object
+			createEvent  event.CreateEvent
+			updateEvent  event.UpdateEvent
+			deleteEvent  event.DeleteEvent
+			genericEvent event.GenericEvent
+		)
+
+		BeforeEach(func() {
+			object = &extensionsv1alpha1.Extension{
+				Spec: extensionsv1alpha1.ExtensionSpec{
+					DefaultSpec: extensionsv1alpha1.DefaultSpec{
+						Type: extensionType,
+					},
+				},
+			}
+			createEvent = event.CreateEvent{
+				Object: object,
+			}
+			updateEvent = event.UpdateEvent{
+				ObjectOld: object,
+				ObjectNew: object,
+			}
+			deleteEvent = event.DeleteEvent{
+				Object: object,
+			}
+			genericEvent = event.GenericEvent{
+				Object: object,
+			}
+		})
+
+		It("should match the type", func() {
+			predicate := HasType(extensionType)
+
+			Expect(predicate.Create(createEvent)).To(BeTrue())
+			Expect(predicate.Update(updateEvent)).To(BeTrue())
+			Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+			Expect(predicate.Generic(genericEvent)).To(BeTrue())
+		})
+
+		It("should not match the type", func() {
+			predicate := HasType("anotherType")
+
+			Expect(predicate.Create(createEvent)).To(BeFalse())
+			Expect(predicate.Update(updateEvent)).To(BeFalse())
+			Expect(predicate.Delete(deleteEvent)).To(BeFalse())
+			Expect(predicate.Generic(genericEvent)).To(BeFalse())
+		})
+	})
+
+	Describe("#HasClass", func() {
+		var (
+			extensionClass *extensionsv1alpha1.ExtensionClass
+
+			object       client.Object
+			createEvent  event.CreateEvent
+			updateEvent  event.UpdateEvent
+			deleteEvent  event.DeleteEvent
+			genericEvent event.GenericEvent
+		)
+
+		JustBeforeEach(func() {
+			object = &extensionsv1alpha1.Extension{
+				Spec: extensionsv1alpha1.ExtensionSpec{
+					DefaultSpec: extensionsv1alpha1.DefaultSpec{
+						Class: extensionClass,
+					},
+				},
+			}
+			createEvent = event.CreateEvent{
+				Object: object,
+			}
+			updateEvent = event.UpdateEvent{
+				ObjectOld: object,
+				ObjectNew: object,
+			}
+			deleteEvent = event.DeleteEvent{
+				Object: object,
+			}
+			genericEvent = event.GenericEvent{
+				Object: object,
+			}
+		})
+
+		testAndVerify := func(classes []extensionsv1alpha1.ExtensionClass, match gomegatypes.GomegaMatcher) {
+			predicate := HasClass(classes...)
+
+			Expect(predicate.Create(createEvent)).To(match)
+			Expect(predicate.Update(updateEvent)).To(match)
+			Expect(predicate.Delete(deleteEvent)).To(match)
+			Expect(predicate.Generic(genericEvent)).To(match)
+		}
+
+		Context("when class is unset", func() {
+			It("should match an empty class (nil)", func() {
+				testAndVerify(nil, BeTrue())
+			})
+
+			It("should match an empty class (empty)", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{""}, BeTrue())
+			})
+
+			It("should match the 'shoot' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"shoot"}, BeTrue())
+			})
+
+			It("should not match the 'garden' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"garden"}, BeFalse())
+			})
+
+			It("should not match multiple classes", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"seed", "garden"}, BeFalse())
+			})
+		})
+
+		Context("when class is set to 'shoot'", func() {
+			BeforeEach(func() {
+				extensionClass = ptr.To[extensionsv1alpha1.ExtensionClass]("shoot")
+			})
+
+			It("should match an empty class (nil)", func() {
+				testAndVerify(nil, BeTrue())
+			})
+
+			It("should match an empty class (empty)", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{""}, BeTrue())
+			})
+
+			It("should match the 'shoot' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"shoot"}, BeTrue())
+			})
+
+			It("should not match the 'garden' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"garden"}, BeFalse())
+			})
+
+			It("should not match multiple classes", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"seed", "garden"}, BeFalse())
+			})
+		})
+
+		Context("when class is set to 'garden'", func() {
+			BeforeEach(func() {
+				extensionClass = ptr.To[extensionsv1alpha1.ExtensionClass]("garden")
+			})
+
+			It("should not match an empty class (nil)", func() {
+				testAndVerify(nil, BeFalse())
+			})
+
+			It("should not match an empty class (empty)", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{""}, BeFalse())
+			})
+
+			It("should not match the 'shoot' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"shoot"}, BeFalse())
+			})
+
+			It("should match the 'garden' class", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"garden"}, BeTrue())
+			})
+
+			It("should not match multiple classes", func() {
+				testAndVerify([]extensionsv1alpha1.ExtensionClass{"shoot", "seed"}, BeFalse())
+			})
+		})
+	})
+
+	Describe("#AddTypePredicate", func() {
+		var (
+			object           client.Object
+			extensionTypeFoo = "foo"
+
+			createEvent   event.CreateEvent
+			updateEvent   event.UpdateEvent
+			deleteEvent   event.DeleteEvent
+			genericEvent  event.GenericEvent
+			truePredicate predicate.Predicate = predicate.NewPredicateFuncs(func(client.Object) bool { return true })
+		)
+
+		BeforeEach(func() {
+			object = &extensionsv1alpha1.Extension{
+				Spec: extensionsv1alpha1.ExtensionSpec{
+					DefaultSpec: extensionsv1alpha1.DefaultSpec{
+						Type: extensionType,
+					},
+				},
+			}
+			createEvent = event.CreateEvent{
+				Object: object,
+			}
+			updateEvent = event.UpdateEvent{
+				ObjectOld: object,
+				ObjectNew: object,
+			}
+			deleteEvent = event.DeleteEvent{
+				Object: object,
+			}
+			genericEvent = event.GenericEvent{
+				Object: object,
+			}
+		})
+
+		It("should add the HasType predicate of the passed extension to the given list of predicates", func() {
+			predicates := AddTypeAndClassPredicates([]predicate.Predicate{truePredicate}, extensionsv1alpha1.ExtensionClassShoot, extensionType)
+
+			Expect(predicates).To(HaveLen(3))
+			Expect(reflect.ValueOf(predicates[2])).To(Equal(reflect.ValueOf(truePredicate)), "predicate list should contain the passed predicate at last element")
+			pred := predicate.And(predicates...)
+
+			Expect(pred.Create(createEvent)).To(BeTrue())
+			Expect(pred.Update(updateEvent)).To(BeTrue())
+			Expect(pred.Delete(deleteEvent)).To(BeTrue())
+			Expect(pred.Generic(genericEvent)).To(BeTrue())
+
+			predicates = AddTypeAndClassPredicates([]predicate.Predicate{truePredicate}, extensionsv1alpha1.ExtensionClassShoot, extensionTypeFoo)
+
+			Expect(predicates).To(HaveLen(3))
+			pred = predicate.And(predicates...)
+
+			Expect(pred.Create(createEvent)).To(BeFalse())
+			Expect(pred.Update(updateEvent)).To(BeFalse())
+			Expect(pred.Delete(deleteEvent)).To(BeFalse())
+			Expect(pred.Generic(genericEvent)).To(BeFalse())
+		})
+
+		It("should add OR of all the HasType predicates for the passed extensions to the given list of predicates", func() {
+			predicates := AddTypeAndClassPredicates([]predicate.Predicate{truePredicate}, extensionsv1alpha1.ExtensionClassShoot, extensionType, extensionTypeFoo)
+
+			Expect(predicates).To(HaveLen(3))
+			pred := predicate.And(predicates...)
+
+			Expect(pred.Create(createEvent)).To(BeTrue())
+			Expect(pred.Update(updateEvent)).To(BeTrue())
+			Expect(pred.Delete(deleteEvent)).To(BeTrue())
+			Expect(pred.Generic(genericEvent)).To(BeTrue())
+
+			// checking HasType(extensionTypeFoo)
+			object = &extensionsv1alpha1.Extension{
+				Spec: extensionsv1alpha1.ExtensionSpec{
+					DefaultSpec: extensionsv1alpha1.DefaultSpec{
+						Type: extensionTypeFoo,
+					},
+				},
+			}
+			createEvent = event.CreateEvent{
+				Object: object,
+			}
+			updateEvent = event.UpdateEvent{
+				ObjectOld: object,
+				ObjectNew: object,
+			}
+			deleteEvent = event.DeleteEvent{
+				Object: object,
+			}
+			genericEvent = event.GenericEvent{
+				Object: object,
+			}
+			Expect(pred.Create(createEvent)).To(BeTrue())
+			Expect(pred.Update(updateEvent)).To(BeTrue())
+			Expect(pred.Delete(deleteEvent)).To(BeTrue())
+			Expect(pred.Generic(genericEvent)).To(BeTrue())
+		})
+	})
+
+	Describe("#HasPurpose", func() {
+		var (
+			object        *extensionsv1alpha1.ControlPlane
+			purposeFoo    = extensionsv1alpha1.Purpose("foo")
+			purposeNormal = extensionsv1alpha1.Normal
+
+			createEvent  event.CreateEvent
+			updateEvent  event.UpdateEvent
+			deleteEvent  event.DeleteEvent
+			genericEvent event.GenericEvent
+		)
+
+		BeforeEach(func() {
+			object = &extensionsv1alpha1.ControlPlane{}
+			createEvent = event.CreateEvent{
+				Object: object,
+			}
+			updateEvent = event.UpdateEvent{
+				ObjectOld: object,
+				ObjectNew: object,
+			}
+			deleteEvent = event.DeleteEvent{
+				Object: object,
+			}
+			genericEvent = event.GenericEvent{
+				Object: object,
+			}
+		})
+
+		It("should return true because purpose is 'normal' and spec is nil", func() {
+			predicate := HasPurpose(purposeNormal)
+
+			Expect(predicate.Create(createEvent)).To(BeTrue())
+			Expect(predicate.Update(updateEvent)).To(BeTrue())
+			Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+			Expect(predicate.Generic(genericEvent)).To(BeTrue())
+		})
+
+		It("should return true because purpose is 'normal' and spec is 'normal'", func() {
+			object.Spec.Purpose = &purposeNormal
+			predicate := HasPurpose(purposeNormal)
+
+			Expect(predicate.Create(createEvent)).To(BeTrue())
+			Expect(predicate.Update(updateEvent)).To(BeTrue())
+			Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+			Expect(predicate.Generic(genericEvent)).To(BeTrue())
+		})
+
+		It("should return false because purpose is not 'normal' and spec is nil", func() {
+			predicate := HasPurpose(purposeFoo)
+
+			Expect(predicate.Create(createEvent)).To(BeFalse())
+			Expect(predicate.Update(updateEvent)).To(BeFalse())
+			Expect(predicate.Delete(deleteEvent)).To(BeFalse())
+			Expect(predicate.Generic(genericEvent)).To(BeFalse())
+		})
+
+		It("should return false because purpose does not match", func() {
+			object.Spec.Purpose = &purposeFoo
+			predicate := HasPurpose(extensionsv1alpha1.Exposure)
+
+			Expect(predicate.Create(createEvent)).To(BeFalse())
+			Expect(predicate.Update(updateEvent)).To(BeFalse())
+			Expect(predicate.Delete(deleteEvent)).To(BeFalse())
+			Expect(predicate.Generic(genericEvent)).To(BeFalse())
+		})
+
+		It("should return true because purpose matches", func() {
+			object.Spec.Purpose = &purposeFoo
+			predicate := HasPurpose(purposeFoo)
+
+			Expect(predicate.Create(createEvent)).To(BeTrue())
+			Expect(predicate.Update(updateEvent)).To(BeTrue())
+			Expect(predicate.Delete(deleteEvent)).To(BeTrue())
+			Expect(predicate.Generic(genericEvent)).To(BeTrue())
 		})
 	})
 })
